refactor: change the positions of CarManager and racingGame
